requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0
requests>=2.31.0
sqlalchemy>=2.0.0
pytest>=7.3.1
mypy>=1.3.0
flake8>=6.0.0
pyyaml>=6.0
feast>=0.30.0
